refactor(ui): extract habit card components into separate files
- Split HabitCard into multiple focused components:
  - HabitCardCalendar for calendar visualization
  - HabitCardHeader for title and metadata
  - HabitCardMetadata for tags and time display
  - HabitCardTitle for habit name display
  - HabitCheckbox for completion tracking
  - HabitProgress for progress visualization
- Move TagDisplayUtils to shared utils directory
- Update imports in related components to use new utils location
- Add comprehensive tests for all new extracted components
- Improve test setup with dependency injection mocks
